[ remove ] log 모듈 제거 - 추천 알고리즘 서비스 모듈과 통합
