--- log/src/main/java/com/link_eleven/msa/log/LogApplication.java
@@ -1,17 +1,0 @@
-package com.link_eleven.msa.log;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableFeignClients
-@EnableJpaAuditing
-public class LogApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(LogApplication.class, args);
-	}
-
-}
