--- log/src/main/java/com/link_eleven/msa/log/infrastructure/configuration/AppConfig.java
@@ -1,14 +1,0 @@
-package com.link_eleven.msa.log.infrastructure.configuration;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class AppConfig {
-
-	@Bean
-	public RestTemplate restTemplate() {
-		return new RestTemplate();
-	}
-}
