--- log/src/main/java/com/link_eleven/msa/log/infrastructure/messaging/FeedLogMessage.java
@@ -1,11 +1,0 @@
-package com.link_eleven.msa.log.infrastructure.messaging;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class FeedLogMessage {
-	private Long userId;
-	private String feedTitle;
-}
