--- log/src/main/java/com/link_eleven/msa/log/libs/exception/CustomException.java
@@ -1,15 +1,0 @@
-package com.link_eleven.msa.log.libs.exception;
-
-import lombok.Getter;
-
-@Getter
-public class CustomException extends RuntimeException {
-
-	private final ErrorCode errorCode;
-	private final String message;
-
-	public CustomException(ErrorCode errorCode) {
-		this.errorCode = errorCode;
-		this.message = errorCode.getMessage();
-	}
-}
