--- log/src/main/resources/application-prod.yml
@@ -1,37 +1,0 @@
-spring:
-  application:
-    name: log-service
-
-  datasource:
-    url: ${LOG_DB_URL}
-    username: ${LOG_DB_USER}
-    password: ${LOG_DB_PASSWORD}
-    driver-class-name: org.postgresql.Driver
-
-  kafka:
-    bootstrap-servers: ${BOOTSTRAP_SERVER}
-    consumer:
-      group-id: ${GROUP_ID}
-      auto-offset-reset: earliest
-      key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
-      value-deserializer: org.apache.kafka.common.serialization.StringDeserializer
-    producer:
-      key-serializer: org.apache.kafka.common.serialization.StringSerializer
-      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
-
-  data:
-    redis:
-      host: ${REDIS_HOST}
-      port: ${REDIS_PORT}
-
-gemini:
-  api:
-    key: ${EXTERNAL_API_KEY}
-
-eureka:
-  client:
-    service-url:
-      defaultZone: ${SERVER_URL}
-
-server:
-  port: ${LOG_PORT}
