--- log/src/test/java/com/link_eleven/msa/log/LogApplicationTests.java
@@ -1,13 +1,0 @@
-package com.link_eleven.msa.log;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class LogApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
