+++ recommendation/src/main/java/com/linkeleven/msa/recommendation/infrastructure/configuration/AppConfig.java
@@ -1,0 +1,14 @@
+package com.linkeleven.msa.recommendation.infrastructure.configuration;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class AppConfig {
+
+	@Bean
+	public RestTemplate restTemplate() {
+		return new RestTemplate();
+	}
+}
