+++ recommendation/src/main/resources/application-dev.yml
@@ -25,8 +25,17 @@ spring:
   kafka:
     bootstrap-servers: ${BOOTSTRAP_SERVER}
     consumer:
-      key-deserializer: org.apache.kafka.common.serialization.StringSerializer
-      value-deserializer: org.apache.kafka.common.serialization.StringSerializer
+      group-id: ${GROUP_ID}
+      auto-offset-reset: earliest
+      key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
+      value-deserializer: org.apache.kafka.common.serialization.StringDeserializer
+    producer:
+      key-serializer: org.apache.kafka.common.serialization.StringSerializer
+      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
+
+gemini:
+  api:
+    key: ${EXTERNAL_API_KEY}
 
 eureka:
   client:
